face_attendance/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.21.0
Pillow>=9.0.0
deepface==0.0.79
tf-keras==2.15.0
tensorflow-cpu==2.15.0
customtkinter==5.2.0
